Rename `inspector/event.ts` to `event-manager.ts`.
