+++ packages/breadboard/src/inspector/event-manager.ts
